fixed the weapon pick up
i think this is finally fixed. The issue was that when I destroyed and then created the new non-interactable weapon the place I added it in the list was incorrect. I had been adding it to the end of the list instaed of the place where the previous one was destroyed.